Add PacketReading/Writing/Vanilla
