+++ jcstress-tests/src/jcstress/java/net/minestom/server/utils/ObjectPoolTest.java
@@ -1,6 +1,7 @@
 package net.minestom.server.utils;
 
 import net.minestom.server.network.NetworkBuffer;
+import net.minestom.server.network.packet.PacketVanilla;
 import org.openjdk.jcstress.annotations.*;
 import org.openjdk.jcstress.infra.results.L_Result;
 
@@ -11,7 +12,7 @@ import static org.openjdk.jcstress.annotations.Expect.ACCEPTABLE;
 @Outcome(id = "2", expect = ACCEPTABLE)
 @State
 public class ObjectPoolTest {
-    private final ObjectPool<NetworkBuffer> pool = PacketUtils.PACKET_POOL;
+    private final ObjectPool<NetworkBuffer> pool = PacketVanilla.PACKET_POOL;
 
     @Actor
     public void actor1() {
