+++ src/main/java/net/minestom/server/MinecraftServer.java
@@ -22,6 +22,7 @@ import net.minestom.server.message.ChatType;
 import net.minestom.server.monitoring.BenchmarkManager;
 import net.minestom.server.network.ConnectionManager;
 import net.minestom.server.network.packet.PacketParser;
+import net.minestom.server.network.packet.client.ClientPacket;
 import net.minestom.server.network.packet.server.common.PluginMessagePacket;
 import net.minestom.server.network.packet.server.play.ServerDifficultyPacket;
 import net.minestom.server.network.socket.Server;
@@ -40,7 +41,6 @@ import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.UnknownNullability;
 
-import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
 
@@ -80,13 +80,8 @@ public final class MinecraftServer implements MinecraftConstants {
 
     @ApiStatus.Internal
     public static ServerProcess updateProcess() {
-        ServerProcess process;
-        try {
-            process = new ServerProcessImpl();
-            serverProcess = process;
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
+        ServerProcess process = new ServerProcessImpl();
+        serverProcess = process;
         return process;
     }
 
@@ -189,7 +184,7 @@ public final class MinecraftServer implements MinecraftConstants {
         return serverProcess.bossBar();
     }
 
-    public static @NotNull PacketParser.Client getPacketParser() {
+    public static @NotNull PacketParser<ClientPacket> getPacketParser() {
         return serverProcess.packetParser();
     }
