+++ src/main/java/net/minestom/server/ServerProcess.java
@@ -14,6 +14,7 @@ import net.minestom.server.listener.manager.PacketListenerManager;
 import net.minestom.server.monitoring.BenchmarkManager;
 import net.minestom.server.network.ConnectionManager;
 import net.minestom.server.network.packet.PacketParser;
+import net.minestom.server.network.packet.client.ClientPacket;
 import net.minestom.server.network.socket.Server;
 import net.minestom.server.recipe.RecipeManager;
 import net.minestom.server.registry.Registries;
@@ -104,7 +105,7 @@ public interface ServerProcess extends Registries, Snapshotable {
      * <p>
      * Can be used if you want to convert a buffer to a client packet object.
      */
-    @NotNull PacketParser.Client packetParser();
+    @NotNull PacketParser<ClientPacket> packetParser();
 
     /**
      * Exposed socket server.
