+++ src/main/java/net/minestom/server/network/packet/server/CachedPacket.java
@@ -3,7 +3,7 @@ package net.minestom.server.network.packet.server;
 import net.minestom.server.ServerFlag;
 import net.minestom.server.network.ConnectionState;
 import net.minestom.server.network.NetworkBuffer;
-import net.minestom.server.utils.PacketUtils;
+import net.minestom.server.network.packet.PacketWriting;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -51,7 +51,7 @@ public final class CachedPacket implements SendablePacket {
         SoftReference<FramedPacket> ref = packet;
         FramedPacket cache;
         if (ref == null || (cache = ref.get()) == null) {
-            cache = PacketUtils.allocateTrimmedPacket(state, packetSupplier.get());
+            cache = PacketWriting.allocateTrimmedPacket(state, packetSupplier.get());
             this.packet = new SoftReference<>(cache);
         }
         return cache;
