+++ src/test/java/net/minestom/server/network/SendablePacketTest.java
@@ -1,10 +1,10 @@
 package net.minestom.server.network;
 
 import net.kyori.adventure.text.Component;
+import net.minestom.server.network.packet.PacketWriting;
 import net.minestom.server.network.packet.server.CachedPacket;
 import net.minestom.server.network.packet.server.LazyPacket;
 import net.minestom.server.network.packet.server.play.SystemChatPacket;
-import net.minestom.server.utils.PacketUtils;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -32,7 +32,7 @@ public class SendablePacketTest {
         var cached = new CachedPacket(packet);
         assertSame(packet, cached.packet(ConnectionState.PLAY));
 
-        var buffer = PacketUtils.allocateTrimmedPacket(ConnectionState.PLAY, packet);
+        var buffer = PacketWriting.allocateTrimmedPacket(ConnectionState.PLAY, packet);
         var cachedBuffer = cached.body(ConnectionState.PLAY);
         assertTrue(NetworkBuffer.equals(buffer.body(), cachedBuffer));
         // May fail in the very unlikely case where soft references are cleared
