+++ src/test/java/net/minestom/server/network/socket/ServerAddressTest.java
@@ -1,6 +1,5 @@
 package net.minestom.server.network.socket;
 
-import net.minestom.server.network.packet.PacketParser;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -19,7 +18,7 @@ public class ServerAddressTest {
         assumeTrue(System.getenv("GITHUB_ACTIONS") == null);
 
         InetSocketAddress address = new InetSocketAddress("localhost", 25565);
-        var server = new Server(new PacketParser.Client());
+        var server = new Server();
         server.init(address);
         assertSame(address, server.socketAddress());
         assertEquals(address.getHostString(), server.getAddress());
@@ -35,7 +34,7 @@ public class ServerAddressTest {
         assumeTrue(System.getenv("GITHUB_ACTIONS") == null);
 
         InetSocketAddress address = new InetSocketAddress("localhost", 0);
-        var server = new Server(new PacketParser.Client());
+        var server = new Server();
         server.init(address);
         assertSame(address, server.socketAddress());
         assertEquals(address.getHostString(), server.getAddress());
@@ -51,7 +50,7 @@ public class ServerAddressTest {
         assumeTrue(System.getenv("GITHUB_ACTIONS") == null);
 
         UnixDomainSocketAddress address = UnixDomainSocketAddress.of("minestom.sock");
-        var server = new Server(new PacketParser.Client());
+        var server = new Server();
         server.init(address);
         assertTrue(Files.exists(address.getPath()));
         assertSame(address, server.socketAddress());
@@ -64,8 +63,8 @@ public class ServerAddressTest {
     }
 
     @Test
-    public void noAddressTest() throws IOException {
-        var server = new Server(new PacketParser.Client());
+    public void noAddressTest() {
+        var server = new Server();
         assertDoesNotThrow(server::stop);
     }
 }
